Also test requests with a custom CA
